requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic[email]
python-dotenv
psycopg2
asyncpg
alembic
pytest
httpx
passlib
bcrypt
python-jose
python-multipart